+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.1.0
 
 - Drop support for Dart SDK versions below 2.0.0
+- Add arcsin, arccos, arctan functions
 - Update examples and documentation
 - Switch to [pendantic](https://pub.dartlang.org/packages/pedantic) analysis
   options